pyproject.toml
[project]
name = "quiz-server"
version = "0.1.0"
description = "Quiz server providing websocket endpoints for participation and managing the quiz game"
authors = [
    {name = "Martin Zeleny", email = "martin.zeleny@volny.cz"},
]
readme = "README.md"
license = {text = "MIT"}

requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.112.1",
    "aioconsole>=0.6.2",
    "ruamel-yaml>=0.18.5",
    "dataclasses>=0.6",
    "logging>=0.4.9.6",
    "quiz-common",
]

[tool.uv.sources]
quiz-common = { git = "https://github.com/quiz-cli/quiz-common.git", rev = "main" }
